requirements.txt
# Updated Requirements for Flask-ShopifyBot

The following dependencies are required for the optimized MCP integration:

# Flask and related
flask>=2.0.0
flask-login>=0.6.2
flask-migrate>=4.0.5
python-dotenv>=1.0.1

# Networking and API
aiohttp>=3.9.1
httpx>=0.28.0
beautifulsoup4>=4.12.2

# Google API integration
google-api-python-client>=2.118.0
google-auth>=2.28.2
google-auth-oauthlib>=1.2.0
google-auth-httplib2>=0.1.1

# Server
uvicorn>=0.29.0
gunicorn>=22.0.0
psycopg2-binary>=2.9.9

# MCP and AI
openai>=1.0.0
mcp==1.9.1
sentence-transformers==2.7.0
openai-agents

# API and utilities
fastapi>=0.104.1
pydantic>=2.5.2
python-multipart>=0.0.6
sse-starlette>=1.6.1
